introduce useVotePreference
