+++ app/components/views/ProposalDetailsPage/helpers/VotePreferenceWrapper/VotePreferenceWrapper.jsx
@@ -1,21 +1,8 @@
-import { useDispatch } from "react-redux";
 import { FormattedMessage as T } from "react-intl";
-import { useState, useCallback, useMemo } from "react";
-import { useMachine } from "@xstate/react";
-import { fetchMachine } from "stateMachines/FetchStateMachine";
 import { StakeyBounceXs } from "indicators";
-import * as gov from "actions/GovernanceActions";
 import styles from "./VotePreferenceWrapper.module.css";
+import { useVotePreference } from "./hooks";
 import VotePreference from "./VotePreference";
-
-const getError = (error) => {
-  if (!error) return;
-  if (typeof error === "string") return error;
-  if (typeof error === "object") {
-    if (error.message) return error.message;
-    return JSON.stringify(error);
-  }
-};
 
 const VotePreferenceWrapper = ({
   viewedProposalDetails,
@@ -23,47 +10,13 @@ const VotePreferenceWrapper = ({
   currentVoteChoice,
   votingComplete
 }) => {
-  // XXX move all hooks/state logic to a custom hook ie. useVotePreference
-  const { eligibleTicketCount } = viewedProposalDetails || {};
-  const [newVoteChoice, setVoteOption] = useState(null);
-
-  const dispatch = useDispatch();
-  const onUpdateVoteChoice = (privatePassphrase) =>
-    dispatch(
-      gov.updateVoteChoice(
-        viewedProposalDetails,
-        newVoteChoice,
-        privatePassphrase
-      )
-    );
-  const [state, send] = useMachine(fetchMachine, {
-    actions: {
-      initial: () => ({}),
-      load: (context, event) => {
-        const { privatePassphrase } = event;
-        if (!newVoteChoice) return;
-        onUpdateVoteChoice(privatePassphrase)
-          .then(() => send("RESOLVE"))
-          .catch((error) => {
-            send({ type: "REJECT", error });
-            setVoteOption(null);
-          });
-      }
-    }
-  });
-
-  const error = useMemo(
-    () => state && state.context && getError(state.context.error),
-    [state]
-  );
-
-  const voteSubmitHandler = useCallback(
-    (privatePassphrase) =>
-      !error
-        ? send({ type: "FETCH", privatePassphrase })
-        : send({ type: "RETRY", privatePassphrase }),
-    [send, error]
-  );
+  const {
+    setVoteOption,
+    newVoteChoice,
+    state,
+    voteSubmitHandler,
+    eligibleTicketCount
+  } = useVotePreference(viewedProposalDetails);
 
   switch (state.value) {
     case "idle":
